+++ packages/react-json-tree/src/JSONArrayNode.tsx
@@ -1,35 +1,30 @@
 import React from 'react';
-import PropTypes from 'prop-types';
 import JSONNestedNode from './JSONNestedNode';
-import { CircularPropsPassedThroughJSONNode } from './types';
+import type { CommonInternalProps } from './types';
 
 // Returns the "n Items" string for this node,
 // generating and caching it if it hasn't been created yet.
-function createItemString(data: any) {
+function createItemString(data: unknown) {
   return `${(data as unknown[]).length} ${
     (data as unknown[]).length !== 1 ? 'items' : 'item'
   }`;
 }
 
-interface Props extends CircularPropsPassedThroughJSONNode {
-  data: any;
+interface Props extends CommonInternalProps {
+  data: unknown;
   nodeType: string;
 }
 
 // Configures <JSONNestedNode> to render an Array
-const JSONArrayNode: React.FunctionComponent<Props> = ({ data, ...props }) => (
-  <JSONNestedNode
-    {...props}
-    data={data}
-    nodeType="Array"
-    nodeTypeIndicator="[]"
-    createItemString={createItemString}
-    expandable={data.length > 0}
-  />
-);
-
-JSONArrayNode.propTypes = {
-  data: PropTypes.array,
-};
-
-export default JSONArrayNode;
+export default function JSONArrayNode({ data, ...props }: Props) {
+  return (
+    <JSONNestedNode
+      {...props}
+      data={data}
+      nodeType="Array"
+      nodeTypeIndicator="[]"
+      createItemString={createItemString}
+      expandable={(data as unknown[]).length > 0}
+    />
+  );
+}