+++ packages/redux-devtools-cli/src/bin/openApp.ts
@@ -1,16 +1,30 @@
 import open from 'open';
 import path from 'path';
+import { fileURLToPath } from 'url';
+import { createRequire } from 'module';
 import spawn from 'cross-spawn';
-import { Options } from '../options';
+import type { Options } from '../options.js';
+
+const require = createRequire(import.meta.url);
 
 export default async function openApp(app: true | string, options: Options) {
   if (app === true || app === 'electron') {
     try {
       const port = options.port ? `--port=${options.port}` : '';
+      const host = options.host ? `--host=${options.host}` : '';
+      const protocol = options.protocol ? `--protocol=${options.protocol}` : '';
+
       // eslint-disable-next-line @typescript-eslint/no-var-requires
-      spawn.sync(require('electron') as string, [
-        path.join(__dirname, '..', '..', 'app'),
+      spawn(require('electron') as string, [
+        path.join(
+          path.dirname(fileURLToPath(import.meta.url)),
+          '..',
+          '..',
+          'app'
+        ),
         port,
+        host,
+        protocol,
       ]);
     } catch (error) {
       /* eslint-disable no-console */
@@ -33,7 +47,7 @@ export default async function openApp(app: true | string, options: Options) {
   }
 
   await open(
-    `http://localhost:${options.port}/`,
+    `${options.protocol}://${options.host ?? 'localhost'}:${options.port}/`,
     app !== 'browser' ? { app: { name: app } } : undefined
   );
 }