+++ packages/redux-devtools-cli/src/getport.ts
@@ -1,6 +1,0 @@
-declare module 'getport' {
-  export default function getport(
-    start: number,
-    callback: (e: Error | undefined, port: number) => void
-  ): void;
-}