+++ packages/redux-devtools-extension/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## 3.2.5
+
+### Patch Changes
+
+- a0716740: Fix types for other exports from `@redux-devtools/extension`.
+
+## 3.2.4
+
+### Patch Changes
+
+- 07456db4: Propagate store enhancer generic type when using composeWithDevTools
+
 ## 3.2.3
 
 ### Patch Changes